requirements.txt
fastapi
uvicorn[standard]
openai-whisper
pydantic
pinecone
sentence-transformers
python-docx
numpy
pathlib
typing-extensions
python-dotenv
pywin32
python-multipart>=0.0.20
pytest
supabase
python-multipart
openai>=1.13.3
google-generativeai>=0.8.3
supabase>=2.0.0
